requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
mcp==1.9.1
aiosqlite==0.19.0
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-mock==3.11.1